requirements.txt
pytest==7.2.0
Appium-Python-Client==1.0.2

selenium~=3.141.0